+++ src/test/java/Debug/DebugPlayer.java
@@ -8,7 +8,7 @@ import SpaceInvaders.utilities.InputController;
 import java.awt.*;
 
 public class DebugPlayer extends Player {
-    public DebugPlayer(Point location, double health, double size, AbstractFactory entities, GameState gameState, InputController inputController) {
+    public DebugPlayer(Point location, int health, double size, AbstractFactory entities, GameState gameState, InputController inputController) {
         super(location, health, size, entities);
     }
